upa-portal/backend/requirements.txt
aiohappyeyeballs==2.6.1 ; python_version >= "3.11" and python_version < "4.0"
aiohttp-retry==2.9.1 ; python_version >= "3.11" and python_version < "4.0"
aiohttp==3.13.3 ; python_version >= "3.11" and python_version < "4.0"
aiosignal==1.4.0 ; python_version >= "3.11" and python_version < "4.0"
alembic==1.14.1 ; python_version >= "3.11" and python_version < "4.0"
amqp==5.3.1 ; python_version >= "3.11" and python_version < "4.0"
anthropic==0.84.0 ; python_version >= "3.11" and python_version < "4.0"
anyio==4.8.0 ; python_version >= "3.11" and python_version < "4.0"
asn1crypto==1.5.1 ; python_version >= "3.11" and python_version < "4.0"
async-timeout==5.0.1 ; python_version >= "3.11" and python_full_version < "3.11.3"
attrs==25.4.0 ; python_version >= "3.11" and python_version < "4.0"
bcrypt==4.2.1 ; python_version >= "3.11" and python_version < "4.0"
beautifulsoup4==4.14.3 ; python_version >= "3.11" and python_version < "4.0"
billiard==4.2.1 ; python_version >= "3.11" and python_version < "4.0"
boto3==1.36.9 ; python_version >= "3.11" and python_version < "4.0"
botocore==1.36.9 ; python_version >= "3.11" and python_version < "4.0"
cbor2==5.8.0 ; python_version >= "3.11" and python_version < "4.0"
celery==5.4.0 ; python_version >= "3.11" and python_version < "4.0"
celery[redis]==5.4.0 ; python_version >= "3.11" and python_version < "4.0"
certifi==2024.12.14 ; python_version >= "3.11" and python_version < "4.0"
cffi==2.0.0 ; python_version >= "3.11" and platform_python_implementation != "PyPy" and python_version < "4.0"
charset-normalizer==3.4.1 ; python_version >= "3.11" and python_version < "4.0"
click-didyoumean==0.3.1 ; python_version >= "3.11" and python_version < "4.0"
click-plugins==1.1.1 ; python_version >= "3.11" and python_version < "4.0"
click-repl==0.3.0 ; python_version >= "3.11" and python_version < "4.0"
click==8.1.8 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and platform_system == "Windows"
cryptography==46.0.5 ; python_version >= "3.11" and python_version < "4.0"
deprecated==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
distro==1.9.0 ; python_version >= "3.11" and python_version < "4.0"
dnspython==2.7.0 ; python_version >= "3.11" and python_version < "4.0"
docstring-parser==0.17.0 ; python_version >= "3.11" and python_version < "4.0"
ecdsa==0.19.0 ; python_version >= "3.11" and python_version < "4.0"
email-validator==2.2.0 ; python_version >= "3.11" and python_version < "4.0"
fastapi-pagination==0.12.34 ; python_version >= "3.11" and python_version < "4.0"
fastapi==0.95.2 ; python_version >= "3.11" and python_version < "4.0"
flower==2.0.1 ; python_version >= "3.11" and python_version < "4.0"
frozenlist==1.8.0 ; python_version >= "3.11" and python_version < "4.0"
google-auth==2.49.0 ; python_version >= "3.11" and python_version < "4.0"
greenlet==3.1.1 ; python_version < "3.14" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version >= "3.11"
h11==0.14.0 ; python_version >= "3.11" and python_version < "4.0"
httpcore==1.0.8 ; python_version >= "3.11" and python_version < "4.0"
httpx==0.28.1 ; python_version >= "3.11" and python_version < "4.0"
humanize==4.11.0 ; python_version >= "3.11" and python_version < "4.0"
idna==3.10 ; python_version >= "3.11" and python_version < "4.0"
jinja2==3.1.5 ; python_version >= "3.11" and python_version < "4.0"
jiter==0.8.2 ; python_version >= "3.11" and python_version < "4.0"
jmespath==1.0.1 ; python_version >= "3.11" and python_version < "4.0"
kombu==5.4.2 ; python_version >= "3.11" and python_version < "4.0"
limits==5.8.0 ; python_version >= "3.11" and python_version < "4.0"
mako==1.3.8 ; python_version >= "3.11" and python_version < "4.0"
markupsafe==3.0.2 ; python_version >= "3.11" and python_version < "4.0"
multidict==6.7.1 ; python_version >= "3.11" and python_version < "4.0"
mypy-extensions==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
mypy==1.14.1 ; python_version >= "3.11" and python_version < "4.0"
openai==1.60.2 ; python_version >= "3.11" and python_version < "4.0"
packaging==24.2 ; python_version >= "3.11" and python_version < "4.0"
passlib[bcrypt]==1.7.4 ; python_version >= "3.11" and python_version < "4.0"
pdfminer-six==20251230 ; python_version >= "3.11" and python_version < "4.0"
pdfplumber==0.11.9 ; python_version >= "3.11" and python_version < "4.0"
pillow==9.5.0 ; python_version >= "3.11" and python_version < "4.0"
prometheus-client==0.21.1 ; python_version >= "3.11" and python_version < "4.0"
prompt-toolkit==3.0.50 ; python_version >= "3.11" and python_version < "4.0"
propcache==0.4.1 ; python_version >= "3.11" and python_version < "4.0"
psycopg2-binary==2.9.10 ; python_version >= "3.11" and python_version < "4.0"
pyasn1-modules==0.4.2 ; python_version >= "3.11" and python_version < "4.0"
pyasn1==0.6.1 ; python_version >= "3.11" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.11" and platform_python_implementation != "PyPy" and python_version < "4.0" and implementation_name != "PyPy"
pydantic==1.10.21 ; python_version >= "3.11" and python_version < "4.0"
pydantic[email]==1.10.21 ; python_version >= "3.11" and python_version < "4.0"
pyjwt==2.11.0 ; python_version >= "3.11" and python_version < "4.0"
pyopenssl==25.3.0 ; python_version >= "3.11" and python_version < "4.0"
pypdfium2==5.6.0 ; python_version >= "3.11" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.11" and python_version < "4.0"
python-jose[cryptography]==3.3.0 ; python_version >= "3.11" and python_version < "4.0"
python-multipart==0.0.6 ; python_version >= "3.11" and python_version < "4.0"
pytz==2024.2 ; python_version >= "3.11" and python_version < "4.0"
redis==5.2.1 ; python_version >= "3.11" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.11" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.11" and python_version < "4"
s3transfer==0.11.2 ; python_version >= "3.11" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.11" and python_version < "4.0"
slowapi==0.1.9 ; python_version >= "3.11" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
soupsieve==2.8.3 ; python_version >= "3.11" and python_version < "4.0"
sqlalchemy==2.0.37 ; python_version >= "3.11" and python_version < "4.0"
sqlalchemy[mypy]==2.0.37 ; python_version >= "3.11" and python_version < "4.0"
starlette==0.27.0 ; python_version >= "3.11" and python_version < "4.0"
tenacity==8.5.0 ; python_version >= "3.11" and python_version < "4.0"
tornado==6.4.2 ; python_version >= "3.11" and python_version < "4.0"
tqdm==4.67.1 ; python_version >= "3.11" and python_version < "4.0"
twilio==9.10.2 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2025.1 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.3.0 ; python_version >= "3.11" and python_version < "4.0"
uvicorn==0.22.0 ; python_version >= "3.11" and python_version < "4.0"
vine==5.1.0 ; python_version >= "3.11" and python_version < "4.0"
wcwidth==0.2.13 ; python_version >= "3.11" and python_version < "4.0"
webauthn==2.7.0 ; python_version >= "3.11" and python_version < "4.0"
wrapt==2.1.2 ; python_version >= "3.11" and python_version < "4.0"
yarl==1.23.0 ; python_version >= "3.11" and python_version < "4.0"
